Enhancing Pipeline Test
