+++ frontend/src/store/slice/Pipeline/Pipeline.test.ts
@@ -1,15 +1,6 @@
-import { Edge } from "reactflow"
-
 import { expect, describe, test, beforeEach } from "@jest/globals"
-import { AsyncThunk } from "@reduxjs/toolkit"
+import { AnyAction } from "@reduxjs/toolkit"
 
-import {
-  AlgorithmNodePostData,
-  EdgeDict,
-  InputNodePostData,
-  NodeDict,
-  RunPostData,
-} from "api/run/Run"
 import {
   run,
   runByCurrentUid,
@@ -19,6 +10,18 @@ import {
 import * as selectors from "store/slice/Pipeline/PipelineSelectors"
 import reducer, { initialState } from "store/slice/Pipeline/PipelineSlice"
 import {
+  createCancelFulfilledAction,
+  createCancelPendingAction,
+  createCancelRejectedAction,
+  createFulfilledAction,
+  createPendingAction,
+  createPendingActionWithRequestId,
+  createRejectedAction,
+  createRunResultFulfilledAction,
+  createRunResultRejectedAction,
+  runPostData,
+} from "store/slice/Pipeline/PipelineTestUtils"
+import {
   NODE_RESULT_STATUS,
   Pipeline,
   PipelineType,
@@ -26,137 +29,13 @@ import {
 } from "store/slice/Pipeline/PipelineType"
 import { isStartedPipeline } from "store/slice/Pipeline/PipelineUtils"
 import { RootState } from "store/store"
-import { ParamMap, ParamType } from "utils/param/ParamType"
-
-const requestId = "FmYmw6sCHA2Ll5JJfPuJN"
-
-// Factory for reusable mock data structures
-const createNodeDict = (): NodeDict => ({
-  node1: {
-    id: "node1",
-    position: { x: 0, y: 0 },
-    data: {
-      type: "input",
-      label: "input1",
-      path: "/path/to/input",
-      fileType: "image",
-    } as InputNodePostData,
-  },
-  node2: {
-    id: "node2",
-    position: { x: 1, y: 1 },
-    data: {
-      path: "/path/to/algorithm",
-      param: {
-        key: {
-          type: "child",
-          value: "",
-          path: "key",
-          children: {
-            key1: {
-              type: "child",
-              value: "value1",
-              path: "key.key1",
-            } as ParamType,
-          },
-        },
-      } as ParamMap,
-    } as AlgorithmNodePostData,
-  },
-})
-
-const createEdgeDict = (): EdgeDict => ({
-  node1: { id: "node1", type: "type1", data: { key: "value1" } } as Edge,
-  node2: { id: "node2", type: "type2", data: { key: "value2" } } as Edge,
-})
-
-const createSnakemakeParam = (): ParamMap => ({
-  config: {
-    type: "parent",
-    children: {
-      cores: { type: "child", value: 4, path: "config.cores" },
-      memory: { type: "child", value: "16GB", path: "config.memory" },
-      restartTimes: { type: "child", value: 3, path: "config.restartTimes" },
-    },
-  },
-})
-
-const createNwbParam = (): ParamMap => ({
-  config: {
-    type: "parent",
-    children: {
-      filePath: {
-        type: "child",
-        value: "/path/to/nwb/file",
-        path: "nwbParam.filePath",
-      },
-      options: {
-        type: "child",
-        value: { option1: true, option2: false },
-        path: "nwbParam.options",
-      },
-    },
-  },
-})
-
-const runPostData: RunPostData = {
-  name: "record test",
-  nodeDict: createNodeDict(),
-  edgeDict: createEdgeDict(),
-  snakemakeParam: createSnakemakeParam(),
-  nwbParam: createNwbParam(),
-  forceRunList: [],
-}
-
-// Helper to create a fulfilled action for a given action type and payload
-const createFulfilledAction = (
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  actionType: AsyncThunk<any, any, any>,
-  payload = "response data",
-) => ({
-  type: actionType.fulfilled.type,
-  meta: {
-    requestId: requestId,
-    requestStatus: "fulfilled",
-    arg: { runPostData },
-  },
-  payload,
-})
-
-const runPendingAction = {
-  type: run.pending.type,
-  meta: {
-    requestId: requestId,
-    requestStatus: "pending",
-  },
-}
-
-const runRejectedAction = {
-  type: run.rejected.type,
-  meta: {
-    requestId: requestId,
-    requestStatus: "rejected",
-    arg: { runPostData },
-  },
-  error: "error message",
-}
-
-const runByCurrentUidPendingAction = {
-  type: runByCurrentUid.pending.type,
-  meta: { requestId: requestId, requestStatus: "pending" },
-}
-
-const pollRunResultPendingAction = {
-  type: pollRunResult.pending.type,
-  meta: { requestId: requestId, requestStatus: "pending" },
-}
 
 describe("Pipeline State Test", () => {
   describe("Pipeline Run", () => {
     test(run.fulfilled.type, () => {
-      const runFulfilledAction = createFulfilledAction(run)
+      const runFulfilledAction = createFulfilledAction(run) as AnyAction
       const targetState = reducer(
-        reducer(initialState, runPendingAction),
+        reducer(initialState, createPendingActionWithRequestId(run)),
         runFulfilledAction,
       )
 
@@ -174,7 +53,10 @@ describe("Pipeline State Test", () => {
     })
 
     test(run.pending.type, () => {
-      const targetState = reducer(initialState, runPendingAction)
+      const targetState = reducer(
+        initialState,
+        createPendingActionWithRequestId(run),
+      )
       const expectState = {
         run: { status: RUN_STATUS.START_PENDING },
         runBtn: 1,
@@ -184,8 +66,8 @@ describe("Pipeline State Test", () => {
 
     test(run.rejected.type, () => {
       const targetState = reducer(
-        reducer(initialState, runPendingAction),
-        runRejectedAction,
+        reducer(initialState, createPendingActionWithRequestId(run)),
+        createRejectedAction(run),
       )
       const expectState = { run: { status: RUN_STATUS.START_ERROR }, runBtn: 1 }
       expect(targetState).toEqual(expectState)
@@ -197,7 +79,7 @@ describe("Pipeline State Test", () => {
       const runByCurrentUidFulfilledAction =
         createFulfilledAction(runByCurrentUid)
       const targetState = reducer(
-        reducer(initialState, runByCurrentUidPendingAction),
+        reducer(initialState, createPendingAction(runByCurrentUid)),
         runByCurrentUidFulfilledAction,
       )
 
@@ -215,7 +97,10 @@ describe("Pipeline State Test", () => {
     })
 
     test(runByCurrentUid.pending.type, () => {
-      const targetState = reducer(initialState, runByCurrentUidPendingAction)
+      const targetState = reducer(
+        initialState,
+        createPendingAction(runByCurrentUid),
+      )
       const expectState = {
         run: { status: RUN_STATUS.START_PENDING },
         runBtn: 1,
@@ -224,18 +109,9 @@ describe("Pipeline State Test", () => {
     })
 
     test(runByCurrentUid.rejected.type, () => {
-      const runByCurrentUidRejectedAction = {
-        type: runByCurrentUid.rejected.type,
-        meta: {
-          requestId: requestId,
-          requestStatus: "rejected",
-          arg: { runPostData },
-        },
-        error: "error message",
-      }
       const targetState = reducer(
-        reducer(initialState, runByCurrentUidPendingAction),
-        runByCurrentUidRejectedAction,
+        reducer(initialState, createPendingAction(runByCurrentUid)),
+        createRejectedAction(runByCurrentUid),
       )
       const expectState = { run: { status: RUN_STATUS.START_ERROR }, runBtn: 1 }
       expect(targetState).toEqual(expectState)
@@ -244,26 +120,6 @@ describe("Pipeline State Test", () => {
 
   describe("Pipeline PollRunResult", () => {
     test(pollRunResult.fulfilled.type, () => {
-      const pollRunResultFulfilledAction = {
-        type: pollRunResult.fulfilled.type,
-        meta: {
-          requestId: requestId,
-          requestStatus: "fulfilled",
-          arg: { uid: "test-uid" },
-        },
-        payload: {
-          node1: {
-            status: "success",
-            message: "Node 1 completed successfully",
-            name: "Node 1",
-            outputPaths: {
-              output1: { path: "/path/to/output1", type: "images" },
-            },
-          },
-          node2: { status: "failed", message: "Node 2 failed", name: "Node 2" },
-        },
-      }
-
       const mockState: Pipeline = {
         ...initialState,
         run: {
@@ -286,8 +142,8 @@ describe("Pipeline State Test", () => {
       }
 
       const targetState = reducer(
-        reducer(mockState, pollRunResultPendingAction),
-        pollRunResultFulfilledAction,
+        reducer(mockState, createPendingAction(pollRunResult)),
+        createRunResultFulfilledAction(pollRunResult),
       )
 
       const expectState = {
@@ -318,18 +174,9 @@ describe("Pipeline State Test", () => {
     })
 
     test(pollRunResult.rejected.type, () => {
-      const pollRunResultRejectedAction = {
-        type: pollRunResult.rejected.type,
-        meta: {
-          requestId: requestId,
-          requestStatus: "rejected",
-          arg: { uid: "test-uid" },
-        },
-        error: "error message",
-      }
       const targetState = reducer(
-        reducer(initialState, pollRunResultPendingAction),
-        pollRunResultRejectedAction,
+        reducer(initialState, createPendingAction(pollRunResult)),
+        createRunResultRejectedAction(pollRunResult),
       )
       const expectState = {
         ...initialState,
@@ -340,37 +187,10 @@ describe("Pipeline State Test", () => {
   })
 
   describe("Pipeline CancelResult", () => {
-    const cancelResultPendingAction = {
-      type: cancelResult.pending.type,
-      meta: { requestId: requestId, requestStatus: "pending" },
-    }
-
-    const cancelResultRejectedAction = {
-      type: cancelResult.rejected.type,
-      meta: {
-        requestId: requestId,
-        requestStatus: "rejected",
-        arg: { uid: "test-uid" },
-      },
-      error: "error message",
-    }
-
     test(cancelResult.fulfilled.type, () => {
-      const cancelResultFulfilledAction = {
-        type: cancelResult.fulfilled.type,
-        meta: {
-          requestId: requestId,
-          requestStatus: "fulfilled",
-          arg: { uid: "test-uid" },
-        },
-        payload: {
-          message: "Cancellation successful",
-        },
-      }
-
       const targetState = reducer(
-        reducer(initialState, cancelResultPendingAction),
-        cancelResultFulfilledAction,
+        reducer(initialState, createCancelPendingAction(cancelResult)),
+        createCancelFulfilledAction(cancelResult),
       )
 
       const expectState = {
@@ -385,7 +205,10 @@ describe("Pipeline State Test", () => {
 
     // Status is not changing when CancelResult is pending at PipelineSlice
     test(cancelResult.pending.type, () => {
-      const targetState = reducer(initialState, cancelResultPendingAction)
+      const targetState = reducer(
+        initialState,
+        createCancelPendingAction(cancelResult),
+      )
       const expectState = {
         run: { status: RUN_STATUS.START_UNINITIALIZED },
         runBtn: 1,
@@ -396,8 +219,8 @@ describe("Pipeline State Test", () => {
     // Status is not changing when CancelResult is rejected at PipelineSlice
     test(cancelResult.rejected.type, () => {
       const targetState = reducer(
-        reducer(initialState, cancelResultPendingAction),
-        cancelResultRejectedAction,
+        reducer(initialState, createCancelPendingAction(cancelResult)),
+        createCancelRejectedAction(cancelResult),
       )
       const expectState = {
         ...initialState,
